:memo: Refactor: 우리 동네 인기 검색어 로직 수정 및 TODO 작성
