+++ src/main/java/site/moamoa/backend/service/post/command/PostCommandService.java
@@ -5,5 +5,7 @@ import site.moamoa.backend.web.dto.base.SimplePostDTO;
 import java.util.List;
 
 public interface PostCommandService {
+
+    List<SimplePostDTO> findByKeyword(Long memberId, String keyword);
     void updatePostViewCount(Long memberId, Long postId);
 }
