+++ src/main/java/site/moamoa/backend/service/post/query/PostQueryServiceImpl.java
@@ -34,31 +34,6 @@ public class PostQueryServiceImpl implements PostQueryService {
 
     private static final String MEMBER_RECENT_KEYWORD_KEY_PREFIX = "member::";
 
-    // queryDSL 적용 후 searchPostsByKeyword()로 바꾸기
-    @Override
-    public List<SimplePostDTO> findByKeyword(Long memberId, String keyword) {
-        try {
-            redisTemplate.opsForZSet()
-                    .add("member::" + memberId, keyword, LocalDateTime.now().toEpochSecond(ZoneOffset.UTC));
-            log.info("searching time : " + LocalDateTime.now().toEpochSecond(ZoneOffset.UTC));
-
-            String town = memberQueryService.findMemberById(memberId).getTown();
-            redisTemplate.opsForZSet().addIfAbsent("town::" + town, keyword,1);
-            redisTemplate.opsForZSet().incrementScore("town::" + town, keyword,1);
-            //redisTemplate.opsForZSet().add("town::" + town, keyword, 1);
-            
-            //log.info("score : " + redisTemplate.opsForZSet().)
-        } catch (Exception e) {
-            System.out.println(e.toString());
-        }
-
-        List<SimplePostDTO> simplePostDTOS = new ArrayList<>();
-        List<Post> posts = postRepository.findByProductNameContaining(keyword);
-        for(Post post : posts) {
-            simplePostDTOS.add(PostConverter.toSimplePostDTO(post));
-        }
-        return simplePostDTOS;
-    }
 
     @Override
     public PostResponseDTO.GetPosts findPostsByNear(Long memberId) {
